requirements.txt
bedrock-agentcore
strands-agents
bedrock-agentcore-starter-toolkit